+++ NEWS.md
@@ -8,6 +8,7 @@
      
 ## Minor changes 
    
+- `get_anova_table()` supports now an object of class `grouped_anova_test`
 - ANOVA table is now correctly returned when `correction = "none"` for repeated measures ANOVA
 - `NAs` are now automatically removed before quantile computation for identifying outliers (@IndrajeetPatil, [#10](https://github.com/kassambara/rstatix/issues/10)).
 - Unquoted factor variable name is now supported in factor manipulation functions: `set_ref_level()`, `reorder_levels()` and `make_valid_levels()`